Implement VolumeSampleProvider to gain in and output
fixup! Implement VolumeSampleProvider to gain in and output

+++ src/blackcore/afv/audio/input.h
@@ -74,8 +74,8 @@ namespace BlackCore
                 int opusBytesEncoded() const;
                 void setOpusBytesEncoded(int opusBytesEncoded);
 
-                float volume() const;
-                void setVolume(float volume);
+                double volume() const;
+                void setVolume(double volume);
 
                 void start(const QAudioDeviceInfo &inputDevice);
                 void stop();
@@ -95,9 +95,9 @@ namespace BlackCore
 
                 bool m_started = false;
                 int m_opusBytesEncoded = 0;
-                float m_volume = 1.0f;
+                double m_volume = 1.0;
                 int m_sampleCount = 0;
-                float m_maxSampleInput = 0;
+                qint16 m_maxSampleInput = 0;
 
                 const int c_sampleCountPerEvent = 4800;
                 const float maxDb = 0;
